Start integration of SDXL

+++ StableDiffusion.ML.OnnxRuntime/UNet.cs
@@ -12,7 +12,7 @@ namespace StableDiffusion.ML.OnnxRuntime
             var input = new List<NamedOnnxValue> {
                 NamedOnnxValue.CreateFromTensor("encoder_hidden_states", encoderHiddenStates),
                 NamedOnnxValue.CreateFromTensor("sample", sample),
-                NamedOnnxValue.CreateFromTensor("timestep", new DenseTensor<long>(new long[] { timeStep }, new int[] { 1 }))
+                NamedOnnxValue.CreateFromTensor("timestep", new DenseTensor<float>(new float[] { timeStep }, new int[] { 1 }))
             };
 
             return input;
